+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+# [1.1.0] - 2025-09-26
+### Added
+
+-   **Added bones to tshirt**
+-   **Added settings panel**
+
 # [1.0.2] - 2025-09-26
 ### Added